backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-chess>=1.999
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
google-genai>=1.0.0
python-dotenv>=1.0.0
